pyproject.toml
[project]
name = "train-tickets-tgbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "orjson>=3.11.1",
    "pre-commit>=4.2.0",
    "pytest-asyncio>=1.1.0",
    "python-telegram-bot>=22.3",
    "structlog>=25.4.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]
